Change method parameters

+++ src/coututil.rs
@@ -1,4 +1,4 @@
-pub fn print_vector(vector: &Vec<i32>, columns: u32) {
+pub fn print_vector(vector: &Vec<i32>, columns: u128) {
     let mut index = 0;
 
     for i in vector {
@@ -18,11 +18,11 @@ pub fn print_vector_len(vector: &Vec<i32>) {
     println!("Vector length: {}", vector.len());
 }
 
-pub fn print_memory_access(count: u32) {
+pub fn print_memory_access(count: u128) {
     println!("Memory accesses count: {}", count);
 }
 
-pub fn print_comparisons(count: u32) {
+pub fn print_comparisons(count: u128) {
     println!("Comparisons count: {}", count);
 }
